asdf

+++ app/controllers/test_controller.rb
@@ -5,13 +5,14 @@ class TestController < ApplicationController
       @results = []
       thisorder = Order.find(368)
       thisorder.quantities.each do |q|
-        samsies = thisorder.quantities.where(product: q.product)
-        newqty = samsies.sum(:qty)
-        @results << 'SAMSIES = ' + samsies.count.to_s
-        @results << 'NEWQTY = ' + newqty.to_s
-        q.update(qty: newqty)
-        others = samsies.where.not(id: samsies.first.id)
-        others.destroy_all
+        if samsies = thisorder.quantities.where(product: q.product)
+          newqty = samsies.sum(:qty)
+          @results << 'SAMSIES = ' + samsies.count.to_s
+          @results << 'NEWQTY = ' + newqty.to_s
+          q.update(qty: newqty)
+          others = samsies.where.not(id: samsies.first.id)
+          others.destroy_all
+        end
       end
 
       # dbh = RDBI.connect :ODBC, :db => "wholesaleportal"
